Rename sources from msdfgen_c to msdfgen-c
